pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.55.0",
    "google-genai>=1.23.0",
    "openai>=1.93.0",
    "gitpython>=3.1.0",
    "requests>=2.31.0",
]
